[REFACTOR] Moved Rest API to `.api` package along with models
