+++ app/src/main/java/dev/hossain/weatheralert/data/WeatherRepository.kt
@@ -1,6 +1,8 @@
 package dev.hossain.weatheralert.data
 import com.slack.eithernet.ApiResult
 import com.squareup.anvil.annotations.ContributesBinding
+import dev.hossain.weatheralert.api.WeatherApi
+import dev.hossain.weatheralert.api.WeatherForecast
 import dev.hossain.weatheralert.db.CityForecast
 import dev.hossain.weatheralert.db.CityForecastDao
 import dev.hossain.weatheralert.di.AppScope
