+++ app/src/test/java/dev/hossain/weatheralert/data/WeatherRepositoryTest.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
 import com.slack.eithernet.ApiResult
+import dev.hossain.weatheralert.api.WeatherApi
 import dev.hossain.weatheralert.db.CityForecastDao
 import dev.hossain.weatheralert.di.DaggerTestAppComponent
 import dev.hossain.weatheralert.di.NetworkModule
